+++ CHANGELOG.md
@@ -16,6 +16,15 @@
 * [MIC-6063] Implement token providing logic for ASDK requests
 * [MIC-6184] Implement SampleTokenProvider in ASDKSample
 * [MIC-6513+MIC-6279] Implementation of PaymentController from v3/root is now added master version. PaymentController - let's you have all the logic needed for payments without depening on UI.
+* [MIC-6582] CustomerKey editing now available in ASDKSample app
+* [MIC-6584] Switching terminal now available in ASDKSample app
+
+## [2.11.2] - 2022-10-13Z
+
+### Fixed
+
+* pull/198 modalViewController.popupStyle now uses configuration popupStyle (bug fix)
+* [MIC-6735] Response validation by condition `success == true && errorCode == 0`
 
 ## [2.11.1] - 2022-10-10Z